Convert WavPack encoder for standalone use.
Overhaul the WavPack residual encoding routine with a (hopefully)
more comprehensive handling of the "holding_one"/"holding_zero" values.
Not yet perfected, however; some test files don't yet decode properly.

Widen some 32-bit calls to write_bits to use 64-bit mode
since they were triggering assertion errors when compiled in 32-bit mode.

+++ src/bitstream_w.c
@@ -204,7 +204,7 @@ write_bits_actual_le(Bitstream* bs, unsigned int count, int value)
     struct bs_callback* callback;
 
     assert(value >= 0);
-    assert(value < (1l << count));
+    assert(value < (int64_t)(1L << count));
 
     while (count > 0) {
         /*chop off up to 8 bits to write at a time*/
@@ -269,7 +269,7 @@ write_bits64_actual_be(Bitstream* bs, unsigned int count, uint64_t value)
     struct bs_callback* callback;
 
     assert(value >= 0l);
-    assert(value < (1l << count));
+    assert(value < (int64_t)(1l << count));
 
     while (count > 0) {
         /*chop off up to 8 bits to write at a time*/
@@ -311,7 +311,7 @@ write_bits64_actual_le(Bitstream* bs, unsigned int count, uint64_t value)
     struct bs_callback* callback;
 
     assert(value >= 0);
-    assert(value < (1l << count));
+    assert(value < (int64_t)(1ll << count));
 
     while (count > 0) {
         /*chop off up to 8 bits to write at a time*/
@@ -506,7 +506,7 @@ void
 write_bits64_accumulator(Bitstream* bs, unsigned int count, uint64_t value)
 {
     assert(value >= 0l);
-    assert(value < (1l << count));
+    assert(value < (int64_t)(1l << count));
     bs->bits_written += count;
 }
